Reconstruct equation.h
